% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{Distances of fires v damage done}
\format{
A data frame with 14 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance away from target}
  \item{DAMAGE}{damage the target took}
  ...
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A dataset containing the distance of fires and the damage it did
}
\keyword{datasets}
