% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{mynbin}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{number of elements in 1 iteration}

\item{iter}{number of iterations to run}

\item{a}{lower bound of uniform distribution}

\item{b}{upper bound of uniform distribution}
}
\value{
a random uniform distribution with theoretical distribution plotted for reference
}
\description{
mynbin
}
\examples{
myclt(n=50,iter=10000,a=5,b=10)
}
