% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{data set}

\item{SPECIES}{species to filter on}
}
\value{
makes a graph, prints data to the console
}
\description{
myddt
}
\examples{
\dontrun{myddt(df = ddt, SPECIES = "CCATFISH")}
}
