% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{mynbin}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{number of trials until the rth success is returned}

\item{r}{determined number of success}

\item{p}{probability of a success per trial}
}
\value{
a negative binomial distribution
}
\description{
mynbin
}
\examples{
mynbin(y=10, p=0.4, r=3)
}
