% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.r
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean value}

\item{sigma}{standard defiation value}

\item{a}{upper bound of lower tail problem}
}
\value{
a normal distribution graph with probability place on it
}
\description{
myncurve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
