% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{myread}
\usage{
myread(dird, csv)
}
\arguments{
\item{dird}{the directory string}

\item{csv}{the CSV file you wish to open}
}
\value{
the CSV data to be stored in a variable when opened
}
\description{
myread
}
\examples{
\dontrun{myread("/Users/cameronbristol/Desktop/OU/MATH4753/Lab4/", "SPRUCE.csv")}
}
